W-15854291: Fixed ignored junits: 1. Downgraded core module parent to previous one. 2. Done refactoring by moving few files from internal to api package since it was not getting loaded by junit and was throwing NoClassDefFoundError.
