+++ src/main/java/org/mule/extension/ftp/api/FileError.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.ftp.internal.error;
+package org.mule.extension.ftp.api;
 
 
 import org.mule.runtime.extension.api.error.ErrorTypeDefinition;
