+++ src/main/java/org/mule/extension/ftp/api/UriUtils.java
@@ -4,10 +4,9 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.ftp.internal.util;
+package org.mule.extension.ftp.api;
 
 
-import org.mule.extension.ftp.internal.exception.IllegalPathException;
 import org.mule.runtime.api.exception.MuleRuntimeException;
 
 import java.net.URI;
