+++ src/main/java/org/mule/extension/ftp/internal/FtpInputStream.java
@@ -20,8 +20,8 @@ import org.mule.extension.ftp.internal.connection.FileSystem;
 import org.mule.extension.ftp.internal.connection.FtpFileSystem;
 import org.mule.extension.ftp.internal.connection.ManagerBasedConnectionSource;
 import org.mule.extension.ftp.internal.connection.StaticConnectionSource;
-import org.mule.extension.ftp.internal.exception.DeletedFileWhileReadException;
-import org.mule.extension.ftp.internal.exception.FileBeingModifiedException;
+import org.mule.extension.ftp.api.DeletedFileWhileReadException;
+import org.mule.extension.ftp.api.FileBeingModifiedException;
 import org.mule.extension.ftp.internal.lock.UriLock;
 import org.mule.extension.ftp.internal.stream.AbstractNonFinalizableFileInputStream;
 import org.mule.extension.ftp.internal.stream.ExceptionInputStream;
