+++ src/main/java/org/mule/extension/ftp/internal/command/FtpWriteCommand.java
@@ -10,15 +10,15 @@ import static java.lang.String.format;
 import static org.mule.extension.ftp.api.FileWriteMode.APPEND;
 import static org.mule.extension.ftp.api.FileWriteMode.CREATE_NEW;
 import static org.mule.extension.ftp.api.FileWriteMode.OVERWRITE;
-import static org.mule.extension.ftp.internal.util.UriUtils.createUri;
-import static org.mule.extension.ftp.internal.util.UriUtils.trimLastFragment;
+import static org.mule.extension.ftp.api.UriUtils.createUri;
+import static org.mule.extension.ftp.api.UriUtils.trimLastFragment;
 import static org.mule.extension.ftp.internal.FtpUtils.normalizePath;
 
 import org.mule.extension.ftp.api.ftp.FtpFileAttributes;
 import org.mule.extension.ftp.api.FileWriteMode;
 import org.mule.extension.ftp.internal.operation.WriteCommand;
-import org.mule.extension.ftp.internal.exception.FileAlreadyExistsException;
-import org.mule.extension.ftp.internal.exception.IllegalPathException;
+import org.mule.extension.ftp.api.FileAlreadyExistsException;
+import org.mule.extension.ftp.api.IllegalPathException;
 import org.mule.extension.ftp.internal.lock.NullUriLock;
 import org.mule.extension.ftp.internal.lock.UriLock;
 import org.mule.extension.ftp.internal.connection.FtpFileSystem;
