+++ src/main/java/org/mule/extension/ftp/internal/connection/FtpFileSystem.java
@@ -8,8 +8,8 @@ package org.mule.extension.ftp.internal.connection;
 
 import static java.lang.String.format;
 import static org.apache.commons.lang3.StringUtils.isBlank;
-import static org.mule.extension.ftp.internal.error.FileError.DISCONNECTED;
-import static org.mule.extension.ftp.internal.util.UriUtils.createUri;
+import static org.mule.extension.ftp.api.FileError.DISCONNECTED;
+import static org.mule.extension.ftp.api.UriUtils.createUri;
 import static org.mule.extension.ftp.internal.FtpUtils.createUrl;
 import static org.mule.extension.ftp.internal.FtpUtils.getReplyCodeErrorMessage;
 import static org.mule.extension.ftp.internal.FtpUtils.normalizePath;
@@ -32,7 +32,7 @@ import org.mule.extension.ftp.internal.operation.WriteCommand;
 import org.mule.extension.ftp.internal.lock.URLPathLock;
 import org.mule.extension.ftp.internal.lock.UriLock;
 import org.mule.extension.ftp.internal.subset.SubsetList;
-import org.mule.extension.ftp.internal.util.UriUtils;
+import org.mule.extension.ftp.api.UriUtils;
 import org.mule.extension.ftp.api.FTPConnectionException;
 import org.mule.extension.ftp.api.ftp.FtpFileAttributes;
 import org.mule.extension.ftp.api.ftp.FtpTransferMode;
