+++ src/main/java/org/mule/extension/ftp/internal/error/provider/FileRenameErrorTypeProvider.java
@@ -15,7 +15,7 @@ import java.util.Set;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.unmodifiableSet;
-import static org.mule.extension.ftp.internal.error.FileError.*;
+import static org.mule.extension.ftp.api.FileError.*;
 
 /**
  * Errors that can be thrown both in the {@link org.mule.extension.ftp.internal.FtpOperations#rename(FileSystem, String, String, boolean)}
