+++ src/main/java/org/mule/extension/ftp/internal/source/OnNewFileCommand.java
@@ -8,7 +8,7 @@ package org.mule.extension.ftp.internal.source;
 
 import static java.lang.String.format;
 
-import org.mule.extension.ftp.internal.exception.IllegalPathException;
+import org.mule.extension.ftp.api.IllegalPathException;
 import org.mule.extension.ftp.internal.command.FtpCommand;
 import org.mule.extension.ftp.internal.connection.FtpFileSystem;
