+++ src/test/java/org/mule/extension/ftp/FtpCopyTestCase.java
@@ -9,22 +9,20 @@ package org.mule.extension.ftp;
 import static java.lang.String.format;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.assertThat;
-import static org.mule.extension.ftp.internal.util.UriUtils.createUri;
+import static org.mule.extension.ftp.api.UriUtils.createUri;
 import static org.mule.extension.ftp.api.FileTestHarness.HELLO_WORLD;
-import static org.mule.extension.ftp.internal.error.FileError.FILE_ALREADY_EXISTS;
-import static org.mule.extension.ftp.internal.error.FileError.ILLEGAL_PATH;
+import static org.mule.extension.ftp.api.FileError.FILE_ALREADY_EXISTS;
+import static org.mule.extension.ftp.api.FileError.ILLEGAL_PATH;
 import static org.mule.extension.ftp.AllureConstants.FtpFeature.FTP_EXTENSION;
 import static org.mule.extension.ftp.internal.FtpUtils.normalizePath;
 
-import org.junit.Ignore;
-import org.mule.extension.ftp.internal.exception.FileAlreadyExistsException;
-import org.mule.extension.ftp.internal.exception.IllegalPathException;
+import org.mule.extension.ftp.api.FileAlreadyExistsException;
+import org.mule.extension.ftp.api.IllegalPathException;
 
 import io.qameta.allure.Feature;
 import org.junit.Test;
 
 @Feature(FTP_EXTENSION)
-@Ignore
 public class FtpCopyTestCase extends CommonFtpConnectorTestCase {
 
   private static final String SOURCE_FILE_NAME = "test.txt";
