+++ src/test/java/org/mule/extension/ftp/FtpDirectoryListenerPollFailureConsistencyFunctionaTestCase.java
@@ -11,7 +11,6 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
 import static org.mule.tck.probe.PollingProber.check;
 
-import org.junit.Ignore;
 import org.mule.extension.ftp.api.ftp.FtpFileAttributes;
 import org.mule.runtime.api.exception.MuleException;
 import org.mule.runtime.core.api.event.CoreEvent;
@@ -27,7 +26,6 @@ import java.util.concurrent.ConcurrentHashMap;
 import io.qameta.allure.Description;
 import org.junit.Test;
 
-@Ignore
 public class FtpDirectoryListenerPollFailureConsistencyFunctionaTestCase extends CommonFtpConnectorTestCase {
 
   private static final String CONTENT = "File Content.";
